backend/requirements.txt
# aiosmtplib==2.0.1
anyio==3.6.2
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
# fastapi-mail==1.2.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.20.9
MarkupSafe==2.1.2
pdfkit==1.0.0
pydantic==1.10.5
PyJWT==2.6.0
pymongo==4.3.3
rapidfuzz==2.13.7
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
python-dotenv==0.21.1
